eval/bare_llms/requirements.txt
PyYAML==6.0.1
python-dotenv==1.0.1
groq==0.5.0
openai==1.23.6
anthropic==0.25.0
